Add user page styles, update routing, and refactor components
- Created a new CSS file for user page styles to enhance UI.
- Updated routing in CardDash component to redirect to '/Users' instead of '/student'.
- Refactored import paths for Toolbar in CoachesTable, EditorTable, and SitesTable components.
- Introduced MultipleSelectCheckmarks component for improved site selection functionality.
- Added a new Toolbar component for better grid management.
- Removed commented-out code from Student component for cleaner codebase.
