+++ src/components/Coaches/CoachesTable.js
@@ -5,7 +5,7 @@ import { getingData_coaches, deleteCoach, insertCoach } from '../../api/api'; //
 import { useState, useEffect } from 'react';
 import { Button, MenuItem, Menu } from '@mui/material';
 import CoachForm from './CoachForm';
-import Toolbar from '../Student/Toolbar';
+import Toolbar from '../Toolbar/Toolbar';
 import { useNotification } from "../Notification/NotificationProvider";
 import { useTranslation } from "react-i18next";
 import { prefixer } from 'stylis';
