+++ src/components/Editor/EditorTable.js
@@ -5,7 +5,7 @@ import { getingData_Editors, deleteEditor, getingData_Places, getingData_Users, 
 import { useState, useEffect } from 'react';
 import { Button, MenuItem, Menu } from '@mui/material';
 import EditorForm from './EditorForms';
-import Toolbar from '../Student/Toolbar';
+import Toolbar from '../Toolbar/Toolbar';
 import { useNotification } from "../Notification/NotificationProvider";
 import { useTranslation } from "react-i18next";
 import { prefixer } from 'stylis';
