+++ src/components/junk/Student/Student.js
